feat: #18 Вынес реализацию соответствий в абстракцию
