+++ engine/src/main/java/com/github/otymko/jos/runtime/context/type/collection/V8Map.java
@@ -6,28 +6,25 @@
 package com.github.otymko.jos.runtime.context.type.collection;
 
 import com.github.otymko.jos.exception.MachineException;
-import com.github.otymko.jos.runtime.context.CollectionIterable;
 import com.github.otymko.jos.runtime.context.ContextClass;
 import com.github.otymko.jos.runtime.context.ContextConstructor;
 import com.github.otymko.jos.runtime.context.ContextMethod;
-import com.github.otymko.jos.runtime.context.ContextValue;
 import com.github.otymko.jos.runtime.context.IValue;
-import com.github.otymko.jos.runtime.context.IndexAccessor;
-import com.github.otymko.jos.runtime.context.IteratorValue;
 import com.github.otymko.jos.runtime.context.type.ValueFactory;
 import com.github.otymko.jos.runtime.machine.info.ContextInfo;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
 @ContextClass(name = "Соответствие", alias = "Map")
-public class V8Map extends ContextValue implements IndexAccessor,
-        CollectionIterable<V8KeyAndValue> {
-
+public class V8Map extends V8BaseMap {
   public static final ContextInfo INFO = ContextInfo.createByClass(V8Map.class);
 
-  private final Map<IValue, IValue> data = new HashMap<>();
+  private V8Map() {
+    // nope
+  }
+
+  @Override
+  public ContextInfo getContextInfo() {
+    return INFO;
+  }
 
   @ContextConstructor
   public static IValue constructor() {
@@ -36,7 +33,7 @@ public class V8Map extends ContextValue implements IndexAccessor,
 
   @ContextConstructor
   public static IValue constructor(IValue fixedMap) {
-    final var rawFixedMap = (V8FixedMap)fixedMap.getRawValue();
+    final var rawFixedMap = (V8FixedMap) fixedMap.getRawValue();
     if (rawFixedMap == null) {
       throw MachineException.invalidArgumentValueException();
     }
@@ -48,16 +45,13 @@ public class V8Map extends ContextValue implements IndexAccessor,
     return result;
   }
 
+  // region ContextMethod
+
   @ContextMethod(name = "Вставить", alias = "Insert")
   public void insert(IValue key, IValue value) {
     final var rawKey = ValueFactory.rawValueOrUndefined(key);
     final var rawValue = ValueFactory.rawValueOrUndefined(value);
     data.put(rawKey, rawValue);
-  }
-
-  @ContextMethod(name = "Количество", alias = "Count")
-  public IValue count() {
-    return ValueFactory.create(data.size());
   }
 
   @ContextMethod(name = "Очистить", alias = "Clear")
@@ -71,28 +65,7 @@ public class V8Map extends ContextValue implements IndexAccessor,
     data.remove(rawKey);
   }
 
-  private Optional<IValue> getInternal(IValue key) {
-    final var rawKey = ValueFactory.rawValueOrUndefined(key);
-    if (data.containsKey(rawKey)) {
-      return Optional.of(data.get(rawKey));
-    }
-    return Optional.empty();
-  }
-
-  @ContextMethod(name = "Получить", alias = "Get")
-  public IValue get(IValue key) {
-    return getInternal(key).orElse(ValueFactory.create());
-  }
-
-  @Override
-  public IteratorValue iterator() {
-    return V8KeyAndValue.iteratorOf(data.entrySet());
-  }
-
-  @Override
-  public IValue getIndexedValue(IValue index) {
-    return getInternal(index).orElse(ValueFactory.create());
-  }
+  // endregion
 
   @Override
   public void setIndexedValue(IValue index, IValue value) {
@@ -100,7 +73,8 @@ public class V8Map extends ContextValue implements IndexAccessor,
   }
 
   @Override
-  public ContextInfo getContextInfo() {
-    return INFO;
+  public IValue getIndexedValue(IValue index) {
+    return getInternal(index).orElse(ValueFactory.create());
   }
+
 }
